+++ creme/linear_model/softmax.py
@@ -63,7 +63,7 @@ class SoftmaxRegression(base.MultiClassifier):
     def __init__(self, optimizer=None, loss=None, l2=0):
         self.optimizers = collections.defaultdict(functools.partial(
             copy.deepcopy,
-            optim.VanillaSGD(0.01) if optimizer is None else optimizer
+            optim.SGD(0.01) if optimizer is None else optimizer
         ))
         self.loss = optim.CrossEntropy() if loss is None else loss
         self.l2 = l2