+++ creme/model_selection.py
@@ -4,6 +4,7 @@ Utilities for measuring the performance of online learning algorithms.
 import math
 
 from . import base
+from . import compose
 from . import stream