+++ creme/multioutput/chain.py
@@ -64,7 +64,7 @@ class ClassifierChain(BaseChain, base.MultiOutputClassifier):
             ...     model = model.fit_one(x, y)
 
             >>> metric
-            Jaccard: 0.304314
+            Jaccard: 0.417059
 
     """
 
@@ -104,7 +104,7 @@ class RegressorChain(BaseChain, base.MultiOutputRegressor):
             >>> model = multioutput.RegressorChain(
             ...     model=(
             ...         preprocessing.StandardScaler() |
-            ...         linear_model.LinearRegression(intercept_lr=0.7)
+            ...         linear_model.LinearRegression(intercept_lr=0.3)
             ...     ),
             ...     order=[0, 1, 2]
             ... )
@@ -112,7 +112,7 @@ class RegressorChain(BaseChain, base.MultiOutputRegressor):
             >>> metric = metrics.RegressionMultiOutput(metrics.MAE())
 
             >>> model_selection.online_score(X_y, model, metric)
-            MAE: 16.170443
+            MAE: 16.970895...
 
     """